Remove `operator!=` now that C++20 lets us use `operator==`.
Also switches to defaulting `operator==` where we can, and adds a CRTP
helper to let us do this even when using CRTP classes. It turns out to
be quite surprising what is needed to allow a CRTP base class to not
interfere in the defaulting of comparison functions, and so I've
extracted it into its own helper and tried to document it. We don't have
a lot of users yet, but it should ensure things like `Printable`
continue to work without interfering with other patterns.