Switch our main macOS actions to Arm.
GitHub has publicly available Arm macOS runners now:
https://github.blog/changelog/2024-01-30-github-actions-introducing-the-new-m1-macos-runner-available-to-open-source/

These runners are used when the OS label is `macos-14` instead of
`macos-12`.

This PR switches the main PR and merge queue macOS runs to that OS. It
adds a push-only `macos-12` run to check post-merge whether Intel macOS
keeps working as we don't want that to break too badly. Given how much
faster the M1 is, it didn't seem worth it to keep this in the critical
path of development. And the Intel path here is more likely than others
to diminish in importance over time.

All our OS predicates in the action encoded a specific OS version which
no longer works given two macOS versions, so I've switch all the OS
predicates to a generic pattern that doesn't encode a version.

Example run on our `actions-test` branch: https://github.com/carbon-language/carbon-lang/actions/runs/7740538377

Note that before landing this, we'll need to turn off the requirement
for macos-12 to test actions on our merge queue, and then after this
runs we'll need to turn back on macos-14.

We can try to change how we name things to avoid this kind of churn if
folks prefer.