Add Inits to blocks to emit initialization insructions for globals
`Inits` block is now static block within a `SemIR` which will be used to
emit initialization instructions for variables in the `Package` scope.

inst_block_stack now has additional methods to handle `Inits` block
separately, this block can be popped without being finalized allowing to
accumulate between all instances of variables.

At the end of the `check` phase, if this block is not empty , the
function `__global_init` will be added with this block being inserted
into it.

This significantly changes the `SemIR` output, that's why this commit
updates a lot of the test cases.